+++ src/test/java/org/sopt/app/common/fixtures/SoptampPointFixture.java
@@ -1,33 +1,11 @@
 package org.sopt.app.common.fixtures;
 
-import static org.sopt.app.common.fixtures.SoptampUserFixture.SOPTAMP_USER_1;
-import static org.sopt.app.common.fixtures.SoptampUserFixture.SOPTAMP_USER_2;
-import static org.sopt.app.common.fixtures.SoptampUserFixture.SOPTAMP_USER_3;
-
 import java.util.List;
 import org.sopt.app.application.soptamp.SoptampPointInfo.PartPoint;
 import org.sopt.app.application.soptamp.SoptampPointInfo.PartRank;
-import org.sopt.app.application.soptamp.SoptampPointInfo.Point;
 import org.sopt.app.domain.enums.Part;
 
 public class SoptampPointFixture {
-
-    public static final Point POINT_1 = Point.builder()
-            .id(1L)
-            .generation(1L)
-            .soptampUserId(SOPTAMP_USER_1.getId())
-            .points(100L).build();
-    public static final Point POINT_2 = Point.builder()
-            .id(2L)
-            .generation(1L)
-            .soptampUserId(SOPTAMP_USER_2.getId())
-            .points(200L).build();
-    public static final Point POINT_3 = Point.builder()
-            .id(3L)
-            .generation(1L)
-            .soptampUserId(SOPTAMP_USER_3.getId())
-            .points(300L).build();
-    public static final List<Point> SOPTAMP_POINT_LIST = List.of(POINT_1, POINT_2, POINT_3);
 
     public static final PartRank PART_RANK_PLAN = PartRank.builder()
             .part(Part.PLAN.getPartName())
